requirements.txt
apache-airflow
llama-index
qdrant-client
beautifulsoup4
requests
python-dotenv
openai
tiktoken
pydantic